removed old versions and replaced them with new ones
